Moving detect cycle function to utils.
Also adding some unit tests.
